requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.11.17
distro==1.9.0
Django==4.1.4
django-bootstrap5==23.4
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openai==1.3.5
pillow==10.2.0
pydantic==2.6.0
pydantic_core==2.16.1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
python-dotenv==1.0.1
